Pantalla y conexion con la api de pagos
