+++ src/pages/PaymentPage/PaymentPage.css
@@ -1,0 +1,17 @@
+.transactionBox {
+    display: flex;
+    justify-content: center;
+    padding-top: 10rem;
+    padding-bottom: 10rem;
+  }
+  
+  .formStyle {
+    padding: 20px;
+    border-radius: 8px;
+    max-width: 600px;
+    width: 100%;
+    box-shadow: 2px 2px 10px gray;
+    background-color: white;
+    display: flex;
+    flex-direction: column;
+  }
